Modified files to pass clang, cppcheck, and cspell tests

+++ src/libaktualizr/primary/reportqueue.cc
@@ -99,7 +99,7 @@ void ReportQueue::flushQueue() {
     } else if (response.http_status_code == 413) {
       if (report_array.size() > 1) {
         // if 413 is received to posting of more than one event then try sending less events next time
-        cur_event_number_limit_ = report_array.size() > 2 ? report_array.size() / 2 : 1;
+        cur_event_number_limit_ = static_cast<int>(report_array.size() > 2 ? report_array.size() / 2 : 1);
       } else {
         LOG_WARNING << "Dropping a report event " << report_array[0].get("id", "unknown") << " since the server `"
                     << config.tls.server << "` cannot digest it (413).";
